requirements.txt
streamlit==1.32.0
supabase==2.4.1
groq==0.5.0
python-dotenv==1.0.1
Pillow==10.2.0
pydantic==2.6.4
pillow-heif==0.15.0
PyMuPDF==1.23.26